fix: HTTP错误返回结构自定义

+++ app/user/service/internal/biz/demo.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Demo struct {
-	Id   int64  `json:"id"`
+	Id   int32  `json:"id"`
 	Name string `json:"name"`
 }
 
 type DemoRepo interface {
-	GetDemo(ctx context.Context, id int64) (*Demo, error)
+	GetDemo(ctx context.Context, id int32) (*Demo, error)
 }
 
 type DemoUseCase struct {
@@ -23,6 +23,6 @@ func NewDemoUseCase(repo DemoRepo, logger log.Logger) *DemoUseCase {
 	return &DemoUseCase{repo: repo, log: log.NewHelper(log.With(logger, "module", "usecase/address"))}
 }
 
-func (d *DemoUseCase) GetDemo(ctx context.Context, id int64) (*Demo, error) {
+func (d *DemoUseCase) GetDemo(ctx context.Context, id int32) (*Demo, error) {
 	return d.repo.GetDemo(ctx, id)
 }
